requirements.txt
requests
beautifulsoup4
markdown2
pdfkit
python-dotenv
google-generativeai
selenium
